move tests into subdirectories, split up functional tests
