+++ src/test/integration/redux.js
@@ -1,5 +1,5 @@
-import { ORM, Session, createReducer, createSelector } from '../';
-import { createTestModels } from './utils';
+import { ORM, Session, createReducer, createSelector } from '../../';
+import { createTestModels } from '../helpers';
 
 describe('Redux integration', () => {
     let orm;
